# backend/api/build/resources/main/openapi/dopashift-api-v1.yml
openapi: 3.0.3
info:
  title: DopaShift API
  description: |
    REST API for the DopaShift multi-goal life-tracking platform.
    
    ## Versioning Policy
    - All endpoints are prefixed with `/v1/`
    - Previous API versions are supported for 6 months after a new version is released.
    - Breaking changes will only be introduced in new major versions.
    
    ## Authentication
    All endpoints (except /v1/auth/*) require a valid Bearer token in the Authorization header.
    Tokens are obtained via the Keycloak-proxied authentication endpoints.
  version: 1.0.0
  contact:
    name: DopaShift Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dopashift.app
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/profile:
    get:
      tags: [Profile]
      operationId: getProfile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
    put:
      tags: [Profile]
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
  /v1/profile/photo:
    post:
      tags: [Profile]
      operationId: uploadPhoto
      summary: Upload profile photo (JPEG/PNG, max 5MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoUploadResponse'
        '400':
          description: Invalid file type or size
    delete:
      tags: [Profile]
      operationId: deletePhoto
      responses:
        '204':
          description: Photo deleted
  /v1/profile/password:
    post:
      tags: [Profile]
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'

  # === LLM Settings ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, correlationId, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Goal name must be 1-100 characters"
        details:
          type: object
          additionalProperties: true
        correlationId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # === Auth ===
...
    UpdateProfileRequest:
      type: object
      properties:
        locale:
          type: string
          enum: [en, hi, mr]
        accentColor:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        timezone:
          type: string
        quietHoursStart:
          type: string
        quietHoursEnd:
          type: string
    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        locale:
          type: string
        accentColor:
          type: string
        timezone:
          type: string
        quietHoursStart:
          type: string
        quietHoursEnd:
          type: string
        photoUrl:
          type: string
    PhotoUploadResponse:
      type: object
      properties:
        photoUrl:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # === LLM ===
...